Rewrite NCPHop with a flagless, stable speed
